+++ CHANGES.rst
@@ -4,6 +4,27 @@ Changes
 Unreleased
 ----------
 
+The exceptions raised by ``dali.frame.Frame()`` and the
+``Frame.pack_len()`` method are changing. Currently, if you pass an
+invalid sequence to initialise a ``Frame`` a ``TypeError`` will be
+raised. In the next release this will change to ``ValueError``. If you
+pass an invalid length to ``pack_len()``, currently a ``ValueError``
+will be raised. In the next release this will change to
+``OverflowError``.
+
+To obtain the new behaviour now, pass ``new_exceptions=True`` to
+``Frame()`` and ``pack_len()``. In the next release, you will be able
+to continue to obtain the current behaviour by passing
+``new_exceptions=False``. In the release after that, passing any value
+for ``new_exceptions`` will issue a warning and in the release after
+*that*, the ``new_exceptions`` keyword argument will be removed.
+
+The new behaviour makes the library more consistent with the
+exceptions raised by the native ``int.from_bytes()`` and
+``int.to_bytes()`` methods.
+
+- Frame exception updates; see above
+
 - Driver updates [jbaptperez, awahlig, rousveiga, rnixx]
 
 0.7.1 (2020-09-02)